Workout Details Screen showing real time graph for calories and steps

+++ app/src/main/java/vytran/cs175/sjsu/alphafitnessandroidapp/UserInfo.java
@@ -5,12 +5,12 @@ package vytran.cs175.sjsu.alphafitnessandroidapp;
  */
 
 public class UserInfo {
-    private long myId;
+    private int myId;
     private String myName;
     private String myGender;
     private float myWeight;
 
-    public UserInfo(long id, String name, String gender, float weight) {
+    public UserInfo(int id, String name, String gender, float weight) {
         myId = id;
         myName = name;
         myGender = gender;
@@ -19,8 +19,8 @@ public class UserInfo {
 
     public UserInfo(){}
 
-    public void setUserId(long Id) { this.myId = Id;}
-    public long getUserId() {return myId;}
+    public void setUserId(int Id) { this.myId = Id;}
+    public int getUserId() {return myId;}
 
     public void setUserName(String name) {myName = name;}
     public String getUserName() {return myName;}
